requirements.txt
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
cryptography==45.0.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
